repo/PhotoEdit/pyproject.toml
[project]
name = "photoedit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow==10.0.1",
]
